frontend/requirements.txt
dash
plotly
requests
python-dotenv